KOF-237: address DeepSource findings in the shell TO code
- topology_shell.cpp: cast the element-matrix size to size_t before the
  multiplication in the two k0.resize() calls, so the product is computed in the
  wider type (bugprone implicit-widening).
- topology_shell_validation.cpp: thread the pass/fail counter through by
  reference instead of a mutable global (matching shell_validation.cpp), and use
  std::array<char, 48> with .data() for the snprintf label buffers instead of a
  C-style array (no C-style arrays, per the engine style rules).

Behaviour unchanged; the native FD sensitivity test still passes.

+++ engine/cpp/topology_shell.cpp
@@ -113,7 +113,7 @@ ShellStiffnessCache build_shell_stiffness_cache(const ShellTopOptInput& in) {
         const auto Ke = kofem::shell::tet_element_stiffness(
             in.vertices, nd[0], nd[1], nd[2], nd[3], in.solid_young, in.solid_poisson);
         std::vector<double>& k0 = cache.k0[e];
-        k0.resize(12 * 12);
+        k0.resize(static_cast<size_t>(12) * 12);
         for (int a = 0; a < 12; ++a)
             for (int b = 0; b < 12; ++b) k0[a * 12 + b] = Ke[a][b];
         std::vector<int>& dofs = cache.dofs[e];
@@ -145,7 +145,7 @@ ShellStiffnessCache build_shell_stiffness_cache(const ShellTopOptInput& in) {
         const auto Ke = kofem::shell::facet_global_stiffness(in.vertices, nd[0], nd[1], nd[2], t,
                                                              in.shell_young, in.shell_poisson);
         std::vector<double>& k0 = cache.k0[e];
-        k0.resize(18 * 18);
+        k0.resize(static_cast<size_t>(18) * 18);
         for (int a = 0; a < 18; ++a)
             for (int b = 0; b < 18; ++b) k0[a * 18 + b] = Ke[a][b];
         std::vector<int>& dofs = cache.dofs[e];
